@@ -1,27 +1,23 @@
-using System;
-using System.Linq;
-using System.Threading.Tasks;
 using SampleSite.Components.Services;
 
-namespace SampleSite.Server.Data
-{
-    public class WeatherForecastService : IWeatherForecastService
-    {
-        private static readonly string[] Summaries = new[]
-        {
-            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-        };
+namespace SampleSite.Server.Data;
 
-        public Task<WeatherForecast[]> GetForecastAsync()
+public class WeatherForecastService : IWeatherForecastService
+{
+    private static readonly string[] Summaries = new[]
+    {
+        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+    };
+
+    public Task<WeatherForecast[]> GetForecastAsync()
+    {
+        var startDate = DateTime.Today;
+        var rng = new Random();
+        return Task.FromResult(Enumerable.Range(1, 5).Select(index => new WeatherForecast
         {
-            var startDate = DateTime.Today;
-            var rng = new Random();
-            return Task.FromResult(Enumerable.Range(1, 5).Select(index => new WeatherForecast
-            {
-                Date = startDate.AddDays(index),
-                TemperatureC = rng.Next(-20, 55),
-                Summary = Summaries[rng.Next(Summaries.Length)]
-            }).ToArray());
-        }
+            Date = startDate.AddDays(index),
+            TemperatureC = rng.Next(-20, 55),
+            Summary = Summaries[rng.Next(Summaries.Length)]
+        }).ToArray());
     }
 }
