@@ -1,7 +1,6 @@
-﻿namespace Toolbelt.Blazor.I18nText.Interfaces
+﻿namespace Toolbelt.Blazor.I18nText.Interfaces;
+
+public interface I18nTextFallbackLanguage
 {
-    public interface I18nTextFallbackLanguage
-    {
-        string FallBackLanguage { get; }
-    }
+    string FallBackLanguage { get; }
 }
