@@ -1,137 +1,129 @@
-﻿using System;
-using System.Reflection;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using System.ComponentModel;
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 using Microsoft.JSInterop;
 
-namespace Toolbelt.Blazor.I18nText.Internals
+namespace Toolbelt.Blazor.I18nText.Internals;
+
+internal class HelperScript : IAsyncDisposable
 {
-    internal class HelperScript
-#if ENABLE_JSMODULE
-        : IAsyncDisposable
-#endif
+    private delegate ValueTask<T> ScriptInvoker<T>(string identifier, params object[] args);
+
+    private bool _Attached = false;
+
+    private readonly SemaphoreSlim _Syncer = new(1, 1);
+
+    private readonly I18nTextOptions _Options;
+
+    private readonly IJSRuntime _JSRuntime;
+
+    private IJSObjectReference? _JSHelper;
+
+    private readonly DotNetObjectReference<HelperScript> _RefThis;
+
+    private readonly ILogger<HelperScript> _Logger;
+
+    private bool? _IsOnline = null;
+
+    private static bool ShowDeveloperGuide() => true;
+
+    public event AsyncEventHandler<I18nTextChangeLanguageEventArgs>? ChangeLanguage;
+
+    public HelperScript(I18nTextOptions options, IJSRuntime jSRuntime, ILogger<HelperScript> logger)
     {
-        private delegate ValueTask<T> ScriptInvoker<T>(string identifier, params object[] args);
+        this._Options = options;
+        this._JSRuntime = jSRuntime;
+        this._Logger = logger;
+        this._RefThis = DotNetObjectReference.Create(this);
+    }
 
-        private bool ScriptLoaded = false;
-
-        private readonly SemaphoreSlim Syncer = new SemaphoreSlim(1, 1);
-
-        private readonly IJSRuntime JSRuntime;
-
-        private static readonly string Namespace = "Toolbelt.Blazor.I18nText";
-
-        private bool? _IsOnline = null;
-
-        public HelperScript(IJSRuntime jSRuntime)
+    [DynamicDependency(DynamicallyAccessedMemberTypes.PublicProperties, typeof(I18nTextOptions))]
+    private async ValueTask<IJSObjectReference?> EnsureScriptAttachedAsync()
+    {
+        if (!this._Attached)
         {
-            this.JSRuntime = jSRuntime;
-        }
-
-        private string GetVersionText()
-        {
-            var assembly = this.GetType().Assembly;
-            var version = assembly
-                .GetCustomAttribute<AssemblyInformationalVersionAttribute>()?
-                .InformationalVersion ?? assembly.GetName().Version!.ToString();
-            return version;
-        }
-
-#if ENABLE_JSMODULE
-        private IJSObjectReference? JSModule = null;
-
-        private async ValueTask<ScriptInvoker<T>?> EnsureScriptEnabledAsync<T>()
-        {
-            if (!this.ScriptLoaded)
+            await this._Syncer.WaitAsync();
+            try
             {
-                await this.Syncer.WaitAsync();
-                try
+                if (!this._Attached)
                 {
-                    if (!this.ScriptLoaded)
+                    var scriptPath = "./_content/Toolbelt.Blazor.I18nText/helper.min.js";
+
+                    // Add version string for refresh token only navigator is online.
+                    // (If the app runs on the offline mode, the module url with query parameters might cause the "resource not found" error.)
+                    try
                     {
-                        var scriptPath = "./_content/" + Namespace + "/script.module.min.js";//?v=" + version;
-
-                        // Add version string for refresh token only navigator is online.
-                        // (If the app runs on the offline mode, the module url with query parameters might cause the "resource not found" error.)
-                        const string moduleScript = "export function isOnLine(){ return navigator.onLine; }";
-                        await using var inlineJsModule = await this.JSRuntime.InvokeAsync<IJSObjectReference>("import", "data:text/javascript;charset=utf-8," + Uri.EscapeDataString(moduleScript));
-                        var isOnLine = await inlineJsModule.InvokeAsync<bool>("isOnLine");
-                        this._IsOnline = isOnLine;
-
-                        if (isOnLine) scriptPath += $"?v={this.GetVersionText()}";
-
-                        this.JSModule = await this.JSRuntime.InvokeAsync<IJSObjectReference>("import", scriptPath);
-                        this.ScriptLoaded = true;
+                        this._IsOnline = await this._JSRuntime.InvokeAsync<bool>("Toolbelt.Blazor.getProperty", "navigator.onLine");
+                        if (this._IsOnline == true) scriptPath += "?v=" + VersionInfo.VersionText;
                     }
-                }
-                catch { }
-                finally { this.Syncer.Release(); }
-            }
-
-            return this.JSModule == null ? default(ScriptInvoker<T>) : this.JSModule.InvokeAsync<T>;
-        }
-#else
-        private async ValueTask<ScriptInvoker<T>?> EnsureScriptEnabledAsync<T>()
-        {
-            if (!this.ScriptLoaded)
-            {
-                await this.Syncer.WaitAsync();
-                try
-                {
-                    if (!this.ScriptLoaded)
+                    catch (JSException e)
                     {
-                        var scriptPath = "./_content/" + Namespace + "/script.min.js";
-
-                        // Add version string for refresh token only navigator is online.
-                        // (If the app runs on the offline mode, the module url with query parameters might cause the "resource not found" error.)
-                        var jsInProcRuntime = this.JSRuntime as IJSInProcessRuntime;
-                        var isOnLine = jsInProcRuntime?.Invoke<bool>("eval", "navigator.onLine") ?? false;
-                        this._IsOnline = isOnLine;
-                        var versionQuery = isOnLine ? $"?v={this.GetVersionText()}" : "";
-
-                        await this.JSRuntime.InvokeVoidAsync("eval", "new Promise(r=>((d,t,s,v)=>(h=>h.querySelector(t+`[src^=\"${s}\"]`)?r():(e=>(e.src=(s+v),e.onload=r,h.appendChild(e)))(d.createElement(t)))(d.head))(document,'script','" + scriptPath + "','" + versionQuery + "'))");
-                        await this.JSRuntime.InvokeVoidAsync("eval", "Toolbelt.Blazor.I18nText.ready");
-                        this.ScriptLoaded = true;
+                        if (ShowDeveloperGuide())
+                        {
+                            if (e.Message.StartsWith("Could not find 'Toolbelt.Blazor.getProperty'"))
+                            {
+                                this._Logger.LogWarning(e.Message);
+                                this._Logger.LogWarning("Please see also: https://github.com/jsakamoto/Toolbelt.Blazor.HotKeys2/issues/17#issuecomment-2195830001");
+                            }
+                            else this._Logger.LogError(e, e.Message);
+                        }
+                        else this._Logger.LogError(e, e.Message);
                     }
+
+                    await using var module = await this._JSRuntime.InvokeAsync<IJSObjectReference>("import", scriptPath);
+                    this._JSHelper = await module.InvokeAsync<IJSObjectReference>("attach", this._RefThis, this._Options);
+                    this._Attached = true;
                 }
-                catch { }
-                finally { this.Syncer.Release(); }
             }
-            return this.ScriptLoaded == false ? default(ScriptInvoker<T>) : this.JSRuntime.InvokeAsync<T>;
-        }
-#endif
-
-        internal static async ValueTask<string> DefaultGetInitialLanguageAsync(IServiceProvider serviceProvider, I18nTextOptions options)
-        {
-            var helperScript = serviceProvider.GetRequiredService<HelperScript>();
-            var invoker = await helperScript.EnsureScriptEnabledAsync<string>();
-            if (invoker == null) return "en-US";
-            return await invoker.Invoke(Namespace + ".initLang", options.PersistanceLevel);
+            catch { }
+            finally { this._Syncer.Release(); }
         }
 
-        internal static async ValueTask DefaultPersistCurrentLanguageAsync(IServiceProvider serviceProvider, string langCode, I18nTextOptions options)
-        {
-            var helperScript = serviceProvider.GetRequiredService<HelperScript>();
-            var invoker = await helperScript.EnsureScriptEnabledAsync<object>();
-            if (invoker == null) return;
-            await invoker.Invoke(Namespace + ".setCurrentLang", langCode, options.PersistanceLevel);
-        }
+        return this._JSHelper;
+    }
 
-        public async ValueTask<bool> IsOnlineAsync()
-        {
-            if (!_IsOnline.HasValue)
-            {
-                await EnsureScriptEnabledAsync<object>();
-            }
-            return _IsOnline!.Value;
-        }
+    internal static async ValueTask<string> DefaultGetInitialLanguageAsync(IServiceProvider serviceProvider, I18nTextOptions options)
+    {
+        var helperScript = serviceProvider.GetRequiredService<HelperScript>();
+        var helper = await helperScript.EnsureScriptAttachedAsync();
+        if (helper == null) return CultureInfo.CurrentUICulture.Name;
+        return await helper.InvokeAsync<string>("getCurrentLang");
+    }
 
-#if ENABLE_JSMODULE
-        public async ValueTask DisposeAsync()
+    internal static async ValueTask DefaultPersistCurrentLanguageAsync(IServiceProvider serviceProvider, string langCode, I18nTextOptions options)
+    {
+        var helperScript = serviceProvider.GetRequiredService<HelperScript>();
+        var helper = await helperScript.EnsureScriptAttachedAsync();
+        if (helper == null) return;
+        await helper.InvokeAsync<string>("setCurrentLang", langCode);
+    }
+
+    public async ValueTask<bool> IsOnlineAsync()
+    {
+        if (!this._IsOnline.HasValue)
         {
-            if (this.JSModule != null) { await this.JSModule.DisposeAsync(); }
+            await this.EnsureScriptAttachedAsync();
         }
-#endif
+        return this._IsOnline!.Value;
+    }
+
+    [JSInvokable(nameof(OnLanguageChanged)), EditorBrowsable(EditorBrowsableState.Never)]
+    public async Task OnLanguageChanged(string lang)
+    {
+        await this.ChangeLanguage.InvokeAsync(this, new(lang));
+    }
+
+    public async ValueTask DisposeAsync()
+    {
+        if (this._JSHelper is not null)
+        {
+            try { await this._JSHelper.InvokeVoidAsync("dispose"); }
+            catch (JSDisconnectedException) { }
+            try { await this._JSHelper.DisposeAsync(); }
+            catch (JSDisconnectedException) { }
+        }
+        this._RefThis.Dispose();
     }
 }
