@@ -38,7 +38,7 @@ else
 @code {
     private Text Text = new Text();
 
-    private WeatherForecast[] forecasts;
+    private WeatherForecast[] forecasts = Array.Empty<WeatherForecast>();
 
     protected override async Task OnInitializedAsync()
     {
