@@ -11,7 +11,7 @@
 @code {
     // Demonstrates how a parent component can supply parameters
     [Parameter]
-    public string Title { get; set; }
+    public string? Title { get; set; }
 
     private Text Text = new Text();
 
