@@ -1,14 +1,11 @@
-﻿using System;
+﻿namespace Toolbelt.Blazor.I18nText;
 
-namespace Toolbelt.Blazor.I18nText
+public class I18nTextChangeLanguageEventArgs : EventArgs
 {
-    public class I18nTextChangeLanguageEventArgs : EventArgs
-    {
-        public string LanguageCode { get; }
+    public string LanguageCode { get; }
 
-        public I18nTextChangeLanguageEventArgs(string langCode)
-        {
-            LanguageCode = langCode;
-        }
+    public I18nTextChangeLanguageEventArgs(string langCode)
+    {
+        this.LanguageCode = langCode;
     }
 }
