@@ -1,7 +1,3 @@
-﻿using System.Collections.Generic;
-using System.Threading.Tasks;
+﻿namespace Toolbelt.Blazor.I18nText.Internals;
 
-namespace Toolbelt.Blazor.I18nText.Internals
-{
-    internal delegate ValueTask<Dictionary<string, string>?> ReadJsonAsTextMapAsync(string jsonUrl, string hash);
-}
+internal delegate ValueTask<Dictionary<string, string>?> ReadJsonAsTextMapAsync(string jsonUrl, string hash);
